Move marshalling tests out of api/types and into lib/services.
