+++ lib/services/github_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2017 Gravitational, Inc.
+Copyright 2021 Gravitational, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,15 +14,22 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package types
+package services
 
 import (
+	"testing"
+
+	"github.com/gravitational/teleport/lib/utils"
 	check "gopkg.in/check.v1"
 )
 
 type GithubSuite struct{}
 
 var _ = check.Suite(&GithubSuite{})
+
+func (r *GithubSuite) SetUpSuite(c *check.C) {
+	utils.InitLoggerForTests(testing.Verbose())
+}
 
 func (s *GithubSuite) TestUnmarshal(c *check.C) {
 	data := []byte(`{"kind": "github",
