+++ lib/services/license_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2018 Gravitational, Inc.
+Copyright 2021 Gravitational, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,12 +14,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package types
+package services
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/gravitational/teleport/lib/utils"
 	"github.com/gravitational/trace"
 	"gopkg.in/check.v1"
 )
@@ -30,6 +31,10 @@ type LicenseSuite struct {
 var _ = check.Suite(&LicenseSuite{})
 var _ = testing.Verbose
 var _ = fmt.Printf
+
+func (r *LicenseSuite) SetUpSuite(c *check.C) {
+	utils.InitLoggerForTests(testing.Verbose())
+}
 
 func (s *LicenseSuite) TestUnmarshal(c *check.C) {
 	type testCase struct {
