+++ lib/services/resetpasswordtoken_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 Gravitational, Inc.
+Copyright 2021 Gravitational, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -13,12 +13,14 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package types
+package services
 
 import (
 	"fmt"
+	"testing"
 	"time"
 
+	"github.com/gravitational/teleport/lib/utils"
 	"gopkg.in/check.v1"
 )
 
@@ -26,6 +28,10 @@ type ResetPasswordTokenSuite struct{}
 
 var _ = check.Suite(&ResetPasswordTokenSuite{})
 var _ = fmt.Printf
+
+func (r *ResetPasswordTokenSuite) SetUpSuite(c *check.C) {
+	utils.InitLoggerForTests(testing.Verbose())
+}
 
 func (s *ResetPasswordTokenSuite) TestUnmarshal(c *check.C) {
 	created, err := time.Parse(time.RFC3339, "2020-01-14T18:52:39.523076855Z")
